refactor: adress review
